backend/requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
joblib==1.5.3
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.2.6
packaging==26.0
pandas==2.3.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.45
starlette==0.50.0
threadpoolctl==3.6.0
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uv==0.10.8
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
boto3==1.28.0
